Update style of pokemon buttons and Modal body
Updated the style of the pokemon buttons by adding the 'btn-primary' class to the pokemon buttons. The template style of the btn-primary was modified by adding a style which increased the width of each button  to occupy 100% of each grid cell for a better look.

Deleted derelict styles which are not needed due to the inherent style capabilities of the bootstrap component btn-primary class